# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: HolydayAPI
  description: |-
    Приложение "Калькулятор отпускных".
  contact:
    email: nirku0soft@yandex.ru
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
  
servers:
  - url: http://localhost:8181



paths:
  /calculacte:
    get:
      description: Получить сумму отпускных по кол-ву дней отпуска или по датам отпуска.
      operationId: getHolidayPay
      parameters:
        - name: salary
          in: query
          required: true
          schema:
            type: number
            format: double

        - name: weekend_day_count
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: vacancy_start
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: vacancy_end
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Получить сумму отпускных
          content:
              application/json:
                schema:
                  type: number
                  format: double
components:
  schemas:   
    eventDTO:
        type: object
        properties:
          id:
            type: integer
            format: int64
            example: 123
          title:
              type: string
              example: Прогулка по реке Сок
          location:
              type: string
              example: село Старосемейкино
          startTime:
              type: string
              format: date-time
              example: 2005-08-09T18:31:42+03
          endTime:
              format: date-time
              type: string
              example: 2005-08-09T18:31:42+03
    ID:
      type: object
      description: Идентификатор
      properties: 
        id:
          type: integer
          format: int64
          example: 123

    userDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        username:
          type: string
          example: fizlrock
        firstname:
          type: string
          example: Никита
        lastname:
          type: string
          example: Кукурузин
        patronymic:
          type: string
          example: Бетонович
        
    
    wrongField:
      type: object
      properties:
        fieldName:
          type: string
          example: user.username
        errorDescripion:
          type: string
          example: To long

  parameters:
    pageNum:
      name: pageNum
      in: query
      description: Номер страницы
      required: true
      schema:
        type: integer
        format: int32
        minimum: 0
    pageSize:
      name: pageSize
      in: query
      description: Максимальное число элементов на странице.
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1


  responses: 
    validationErrorResponse:
      description: Переданы недопустимые параметры
      content:
        application/json:
          schema: 
            type: array
            items:
              $ref: '#/components/schemas/wrongField'

    entityNotFoundResponse:
      description: Сущность с таким ID не найдена

            
                 
        
    